Do not populate EmberENV with default values
This is to enable the Ember runtime to distinguish between `true` by default and explicitly `true`. Which is required to correctly implement the `jQuery.originalEvent` deprecation, where this should only be triggered when jQuery is enabled by default, and not when it is explicitly enabled. Previously `EmberENV['_JQUERY_INTEGRATION']` would be `true` for both cases, with no way to distinguish. Now when the flag is not explicitly set `EmberENV['_JQUERY_INTEGRATION']` will be `undefined`.

The default value will now only be used for the `isFeatureEnabled()` check, but not for populating `EmberENV` anymore.

Fixes https://github.com/emberjs/ember.js/issues/16849