Add `--run-codemod` commandline option.
This allows tools like `@ember/octanify` to pass `--run-codemods` in
it's own CI system without having to worry about properly mocking
`STDIN`.

This is not _really_ expected to be used as an actual flag users would
specify, but it is fine if they do (which is why the `--run-codemod`
flag has a description so that it will show up in `--help` output).